requirements.txt
paddlepaddle>=2.6.0
paddleocr>=2.7.0
transformers>=4.36.0
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
rapidfuzz>=3.5.0
pydantic>=2.5.0
accelerate>=0.25.0
